feat : 라우팅 구현 및 불필요한 파일 제거
