+++ src/app/calendar/page.tsx
@@ -1,7 +1,8 @@
 "use client";
-import React, { useState } from "react";
+import { useState } from "react";
 
 const WEEK = ["일", "월", "화", "수", "목", "금", "토"];
+
 export const showTodayDate = () => {
   const now = new Date();
   const curYear = now.getFullYear();
@@ -20,22 +21,24 @@ interface CalendarHeaderProps {
   onNextMonth: () => void;
 }
 
-const CalendarHeader: React.FC<CalendarHeaderProps> = ({
+function CalendarHeader({
   year,
   month,
   onPrevMonth,
   onNextMonth,
-}) => (
-  <div className="flex justify-between items-center mb-4 p-20">
-    <button onClick={onPrevMonth} className="text-30 font-semibold">
-      &lt;
-    </button>
-    <div className="text-30 font-bold">{`${year}.${String(month).padStart(2, "0")}`}</div>
-    <button onClick={onNextMonth} className="text-30 font-semibold">
-      &gt;
-    </button>
-  </div>
-);
+}: CalendarHeaderProps) {
+  return (
+    <div className="flex justify-between items-center mb-4 p-20">
+      <button onClick={onPrevMonth} className="text-30 font-semibold">
+        &lt;
+      </button>
+      <div className="text-30 font-bold">{`${year}.${String(month).padStart(2, "0")}`}</div>
+      <button onClick={onNextMonth} className="text-30 font-semibold">
+        &gt;
+      </button>
+    </div>
+  );
+}
 
 interface CalendarDayProps {
   day: number;
@@ -44,7 +47,7 @@ interface CalendarDayProps {
 
 const CalendarDay: React.FC<CalendarDayProps> = ({ day, onClick }) => (
   <div
-    className={`p-2 h-100 font-bold text-30 text-center relative cursor-pointer`}
+    className={`p-2 h-150 font-bold text-30 text-center relative cursor-pointer`}
     onClick={onClick}
   >
     {day}
@@ -57,7 +60,6 @@ function Calendar() {
   const [month, setMonth] = useState<number>(curMonth);
   const [day, setDay] = useState<number>(0);
 
-  // Dummy data for demonstration
   const daysInMonth = new Date(year, month, 0).getDate();
   const dayOfWeek = new Date(`${year}-${month}-01`).getDay();
   const allDays = Array.from({ length: daysInMonth }, (_, v) => v + 1);
@@ -74,7 +76,7 @@ function Calendar() {
   };
 
   return (
-    <div className=" w-[1000px] h-[800px]  mx-auto p-4 bg-gray-100 rounded-lg shadow-md">
+    <div className=" w-[1000px] mx-auto p-4 bg-gray-100 rounded-lg shadow-md">
       <CalendarHeader
         year={year}
         month={month}
@@ -84,7 +86,10 @@ function Calendar() {
 
       <div className="grid grid-cols-7 gap-2">
         {WEEK.map((day) => (
-          <div key={day} className="h-50 text-center font-bold text-30">
+          <div
+            key={day}
+            className="flex items-center justify-center h-50 text-center font-bold text-30"
+          >
             {day}
           </div>
         ))}
