+++ src/app/home/page.tsx
@@ -1,0 +1,7 @@
+export default function HomePage() {
+  return (
+    <div className="h-full w-screen min-h-screen justify-center flex items-center bg-gray-100 gap-30 pr-40 text-30">
+      <p className="text-40 font-bold mb-8 text-center">홈페이지</p>
+    </div>
+  );
+}
